// frontend/src/shared/api/generated/schemas/projectProviderAccessResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Bab API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectProviderAccessResponse {
  id: string;
  org_id: string;
  project_id: string;
  provider_id: string;
  allowed_models: string[] | null;
  created_at: string;
  updated_at: string;
}